app/requirements.txt
djangorestframework==3.14.0
requests
celery
Django==3.2
psycopg2-binary
redis